requirements.txt
annotated-types==0.7.0
anyio==4.7.0
boto3==1.35.87
botocore==1.35.87
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
distro==1.9.0
fastapi==0.115.6
git-filter-repo==2.47.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
jmespath==1.0.1
openai==1.58.1
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.4
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
